requirements.txt

pandas==2.1.0
numpy==1.24.3
psycopg2-binary==2.9.7
sqlalchemy==2.0.20
python-dotenv==1.0.0
kaggle==1.5.16
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2